Consistent fonts
Switch to using free OSS fonts hosted on fonts.google.com for
the title and text instead of using the browser default fantasy
font which does not provide consistent branding as it varies
across browsers.

For the title use 'Special Elite' as it looks like typewriter
text and could resonably be stamped into metal.

For the other text use Montserrat which is a nice readable sans
font.

Using these fonts has the side effect of having the website be
somewhat tracked by google as they now get font referrals.